Fixed IPV6 parsing, added application to run all tests by linking to static test libs.

+++ source/components/core/export/core/Util.h
@@ -13,7 +13,7 @@ class ByteArray;
 namespace Util
 {
 
-int NextPowerOfTwo(int value);
+size_t NextPowerOfTwo(size_t value);
 
 //template<class T> T BitsToBytes(T numBits) {
 //    return (numBits + 7) / 8;
